@@ -29,8 +29,6 @@
     "compile_grammar": "npx lezer-generator --output src/Parser/doenet.js src/Parser/doenet.grammar"
   },
   "dependencies": {
-    "@blueprintjs/core": "^4.0.0-beta.0",
-    "@blueprintjs/datetime": "^4.0.0-beta.0",
     "@codemirror/basic-setup": "^0.19.0",
     "@codemirror/commands": "^0.19.5",
     "@codemirror/gutter": "^0.19.9",
@@ -57,7 +55,6 @@
     "crypto-js": "^3.3.0",
     "cssesc": "^3.0.0",
     "csv-parse": "^4.16.3",
-    "cypress-parallel": "^0.12.0",
     "esm-seedrandom": "^3.0.5",
     "handsontable": "^12.1.2",
     "hi-base32": "^0.5.1",
@@ -70,17 +67,17 @@
     "lorem-ipsum": "^2.0.4",
     "math-expressions": "^2.0.0-alpha59",
     "nanoid": "^3.1.31",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.0.3",
     "react-datetime": "^3.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^11.4.2",
     "react-is": "^17.0.1",
-    "react-mathquill": "^1.0.1",
+    "react-mathquill": "^1.0.3",
     "react-measure": "^2.5.2",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
-    "react-select": "^4.3.1",
+    "react-select": "^5.7.0",
     "react-table": "^7.7.0",
     "react-use-measure": "^2.1.1",
     "react-visibility-sensor-v2": "^1.0.0",
@@ -103,6 +100,7 @@
     "@types/react": "^18.0.5",
     "@web/test-runner": "^0.12.19",
     "chai": "^4.2.0",
+    "cypress-parallel": "^0.12.0",
     "esbuild": "^0.14.38",
     "eslint": "^8.14.0",
     "eslint-config-prettier": "^8.2.0",
